Move NativeLoader to JNA package
