+++ src/main/java/io/github/libjsdl/api/SdlInternalInit.java
@@ -1,6 +1,6 @@
 package io.github.libjsdl.api;
 
-import io.github.libjsdl.loader.NativeLoader;
+import io.github.libjsdl.jna.NativeLoader;
 
 final class SdlInternalInit {
