+++ src/main/java/io/github/libjsdl/api/error/SdlError.java
@@ -1,6 +1,6 @@
 package io.github.libjsdl.api.error;
 
-import io.github.libjsdl.loader.NativeLoader;
+import io.github.libjsdl.jna.NativeLoader;
 
 public final class SdlError {
