+++ src/main/java/io/github/libjsdl/api/mouse/SdlMouse.java
@@ -5,7 +5,7 @@ import com.sun.jna.ptr.IntByReference;
 
 import io.github.libjsdl.api.surface.SDL_Surface;
 import io.github.libjsdl.api.video.SDL_Window;
-import io.github.libjsdl.loader.NativeLoader;
+import io.github.libjsdl.jna.NativeLoader;
 
 @SuppressWarnings("checkstyle:AbbreviationAsWordInName")
 public final class SdlMouse {
