+++ src/main/java/io/github/libjsdl/api/version/SdlVersion.java
@@ -1,7 +1,7 @@
 package io.github.libjsdl.api.version;
 
 import io.github.libjsdl.api.log.SDL_version;
-import io.github.libjsdl.loader.NativeLoader;
+import io.github.libjsdl.jna.NativeLoader;
 
 @SuppressWarnings({
         "checkstyle:MagicNumber",
